[Tests] Remove pointless self-test
The imported tests were getting hoisted, so the self-test message was 
the last thing showing in the output. By that point is totally 
redundant. I can get around this by awaiting dynamic imports of the 
other test files, but jshint doesn't like this. In the end, might as 
well just ditch the "Test script: is running" message entirely.